refactor: restructure ai mention module
