+++ packages/core/src/ai-mentions/index.ts
@@ -1,0 +1,11 @@
+export { watchAiMentions } from './watcher';
+export { defaultPromptBuilder } from './prompt';
+export type {
+  AiMentionContext,
+  AiMentionReply,
+  AiMentionSource,
+  AiMentionWatcherHandle,
+  BuildAiMentionPrompt,
+  BuildAiMentionReplyBody,
+  WatchAiMentionsOptions,
+} from './types';
